requirements.txt
python-telegram-bot[job-queue]>=20.0
httpx>=0.23.0